Fetching Users Manually
Caution

Manually dispatching an RTKQ request thunk will create a subscription entry, but it's then up to you to unsubscribe from that data later - otherwise the data stays in the cache permanently. In this case, we always need user data, so we can skip unsubscribing.
Removing a subscription: https://redux-toolkit.js.org/rtk-query/usage/usage-without-react-hooks#removing-a-subscription